Added ExceptionController
Refactor to include all the exceptionHandlers in a controller
