+++ src/main/java/com/attornatus/gerenciamentoDePessoas/controller/EnderecoController.java
@@ -1,16 +1,13 @@
 package com.attornatus.gerenciamentoDePessoas.controller;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -48,19 +45,4 @@ public class EnderecoController {
 
         return new ResponseEntity<Endereco>(endereco, HttpStatus.OK);
     }
-
-    @ExceptionHandler(NoSuchElementException.class)
-	public ResponseEntity<String> noSuchElementExceptionHandler() {
-        return new ResponseEntity<String>("Não há registros com a matrícula informada", HttpStatus.BAD_REQUEST);
-	}
-
-    @ExceptionHandler(HttpMessageNotReadableException.class)
-	public ResponseEntity<String> httpMessageNotReadableExceptionHandler() {
-        return new ResponseEntity<String>("Há irregularidade no formato JSON", HttpStatus.BAD_REQUEST);
-	}
-
-    @ExceptionHandler(IllegalArgumentException.class)
-	public ResponseEntity<String> illegalArgumentExceptionHandler() {
-        return new ResponseEntity<String>("Mátricula inválida!", HttpStatus.BAD_REQUEST);
-	}
 }
