+++ src/main/java/com/attornatus/gerenciamentoDePessoas/controller/PessoaController.java
@@ -1,18 +1,14 @@
 package com.attornatus.gerenciamentoDePessoas.controller;
 
-import java.time.format.DateTimeParseException;
 import java.util.List;
-import java.util.NoSuchElementException;
 
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -66,24 +62,4 @@ public class PessoaController {
 
         return new ResponseEntity<String>("Pessoa deletada!", HttpStatus.OK);
     }
-
-    @ExceptionHandler(DateTimeParseException.class)
-	public ResponseEntity<String> DateTimeParseExceptionHandler() {
-        return new ResponseEntity<String>("A data deve ser informada no padrão dd/MM/yyyy", HttpStatus.BAD_REQUEST);
-	}
-    
-    @ExceptionHandler(NoSuchElementException.class)
-	public ResponseEntity<String> noSuchElementExceptionHandler() {
-        return new ResponseEntity<String>("Não há registros com a matrícula informada", HttpStatus.BAD_REQUEST);
-	}
-
-    @ExceptionHandler(IllegalArgumentException.class)
-	public ResponseEntity<String> illegalArgumentExceptionHandler() {
-        return new ResponseEntity<String>("Mátricula inválida!", HttpStatus.BAD_REQUEST);
-	}
-
-    @ExceptionHandler(HttpMessageNotReadableException.class)
-	public ResponseEntity<String> httpMessageNotReadableExceptionHandler() {
-        return new ResponseEntity<String>("Há irregularidade no formato JSON", HttpStatus.BAD_REQUEST);
-	}
 }
